services/shared/requirements.txt
sqlalchemy>=2.0.0,<3.0.0
psycopg2-binary>=2.9.0,<3.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
